Add Rancher Local Path Provisioner
StorageOS has become a giant pain to work with. It now
requires a license that is difficult to add even after requesting
it, and the newly required etcd operator will fail on restart,
requiring the entire StorageOS stack to be un/reinstalled (along
with the license) before it will work again.

As such, this update removes the StorageOS installation as the
default option (but leaves the task in place if desired) and instead
uses the Rancher Local Path Provisioner to set up a basic storage
solution. This update also makes the resulting 'local-path' storageclass
the default in Kubernetes, so the 'default' storage type configuration
setting in the PostgreSQL Operator can be used without any additional
updates.

For more information about the provisioner, please see:

https://github.com/rancher/local-path-provisioner